Refactor Sampler node
Fixed issues when making a copy of a Sampler node and changing the new one's file:

- Premature free of the original node's sample buffer in the backend
- Premature "cleanup" of the sound datablock in Blender (disabling fake_user)

The sound datablock is now referenced "the proper way" with a PointerProperty
and managed through the native template_ID layout.

Next I'd like to research if we could use aud/audaspace directly and support
audio formats other than WAV and also not having to load the whole file into memory.
Playback directly from disk could also be in its own node with different features.